Move ExpirationType to expiration package
