+++ src/main/java/com/bukrain/shareasy/expiration/ExpirationType.java
@@ -1,4 +1,4 @@
-package com.bukrain.shareasy.blob;
+package com.bukrain.shareasy.expiration;
 
 public enum ExpirationType {
     SINGLE_USE,
