+++ src/main/java/com/bukrain/shareasy/webapi/blob/dto/BlobCreate.java
@@ -1,6 +1,6 @@
 package com.bukrain.shareasy.webapi.blob.dto;
 
-import com.bukrain.shareasy.blob.ExpirationType;
+import com.bukrain.shareasy.expiration.ExpirationType;
 
 public record BlobCreate(String name, ExpirationType expirationType, int size, Integer expire) {
 }
