+++ src/main/java/com/bukrain/shareasy/webapi/blob/dto/BlobUpdate.java
@@ -1,6 +1,6 @@
 package com.bukrain.shareasy.webapi.blob.dto;
 
-import com.bukrain.shareasy.blob.ExpirationType;
+import com.bukrain.shareasy.expiration.ExpirationType;
 
 public record BlobUpdate(ExpirationType expirationType, int expire) {
 }
