+++ src/main/java/com/bukrain/shareasy/webapi/blob/dto/TokenCreate.java
@@ -1,6 +1,6 @@
 package com.bukrain.shareasy.webapi.blob.dto;
 
-import com.bukrain.shareasy.blob.ExpirationType;
+import com.bukrain.shareasy.expiration.ExpirationType;
 
 public record TokenCreate(Integer expire, ExpirationType expirationType) {
 }
